Change supersmoother to new utils code
