+++ supersmoother/utils.py
@@ -72,95 +72,279 @@ def validate_inputs(*arrays, **kwargs):
     return arrays
 
 
-def windowed_sum(*arrays, **kwargs):
-    """Compute the windowed sum of the given arrays
+def windowed_sum_slow(arrays, span, t=None, indices=None, tpowers=0,
+                      period=None, subtract_mid=False):
+    """Compute the windowed sum of the given arrays.
+
+    This is a slow function, used primarily for testing and validation
+    of the faster version of ``windowed_sum()``
 
     Parameters
     ----------
-    *args : array_like
-        Each additional argument is an array that will be windowed
+    arrays : tuple of arrays
+        arrays to window
     span : int or array of ints
-        The span to use for the sum at each point
+        The span to use for the sum at each point. If array is provided,
+        it must be broadcastable with ``indices``
+    indices : array
+        the indices of the center of the desired windows. If ``None``,
+        the indices are assumed to be ``range(len(arrays[0]))`` though
+        these are not actually instantiated.
+    t : array (optional)
+        Times associated with the arrays
+    tpowers : list (optional)
+        Powers of t for each array sum
+    period : float (optional)
+        Period to use, if times are periodic. If supplied, input times
+        must be sorted according to t % period!
     subtract_mid : boolean
         If true, then subtract the middle value from each sum
-    slow : boolean
-        If true, use a slow method to compute the result. This is
-        primarily useful for testing and validation.
-    indices : array
-        the indices of the center of the desired windows. If provided,
-        it must be broadcastable with the span.
 
     Returns
     -------
     arrays : tuple of ndarrays
         arrays containing the windowed sum of each input array
     """
-    span = kwargs.pop('span', None)
-    subtract_mid = kwargs.pop('subtract_mid', False)
-    slow = kwargs.pop('slow', False)
-    indices = kwargs.pop('indices', None)
-
-    if kwargs:
-        raise ValueError("Unrecognized keywords: {0}".format(kwargs.keys()))
-
-    if span is None:
-        raise ValueError("Must provide a positive integer span")
-
     span = np.asarray(span, dtype=int)
     if not np.all(span > 0):
         raise ValueError("span values must be positive")
+    
+    arrays = tuple(map(np.asarray, arrays))
+    N = arrays[0].size
+    if not all(a.shape == (N,) for a in arrays):
+        raise ValueError("sizes of provided arrays must match")
+    
+    t_input = t
+    if t is not None:
+        t = np.asarray(t)
+        if t.shape != (N,):
+            raise ValueError("shape of t must match shape of arrays")
+    else:
+        t = np.ones(N)
+    
+    tpowers = tpowers + np.zeros(len(arrays))
+    if len(tpowers) != len(arrays):
+        raise ValueError("tpowers must be broadcastable with number of arrays")
 
+    if period:
+        if t_input is None:
+            raise ValueError("periodic requires t to be provided")
+        t = t % period
+    
+    if indices is None:
+        indices = np.arange(N)
+    spans, indices = np.broadcast_arrays(span, indices)
+    
+    results = []
+    for tpower, array in zip(tpowers, arrays):
+        if period:
+            result = [sum(array[j % N]
+                          * (t[j % N] + (j // N) * period) ** tpower
+                          for j in range(i - s // 2,
+                                         i - s // 2 + s))
+                      for i, s in np.broadcast(indices, spans)]
+        else:
+            result = [sum(array[j]
+                          * t[j] ** tpower
+                          for j in range(max(0, i - s // 2),
+                                         min(N, i - s // 2 + s)))
+                      for i, s in np.broadcast(indices, spans)]
+        results.append(np.asarray(result))
+
+    if subtract_mid:
+        results = [r - a[indices] * t[indices] ** tp
+                   for r, a, tp in zip(results, arrays, tpowers)]
+        
+    return tuple(results)
+
+
+def windowed_sum(arrays, span, t=None, indices=None, tpowers=0,
+                 period=None, subtract_mid=False):
+    """Compute the windowed sum of the given arrays.
+
+    Parameters
+    ----------
+    arrays : tuple of arrays
+        arrays to window
+    span : int or array of ints
+        The span to use for the sum at each point. If array is provided,
+        it must be broadcastable with ``indices``
+    indices : array
+        the indices of the center of the desired windows. If ``None``,
+        the indices are assumed to be ``range(len(arrays[0]))`` though
+        these are not actually instantiated.
+    t : array (optional)
+        Times associated with the arrays
+    tpowers : list (optional)
+        Powers of t for each array sum
+    period : float (optional)
+        Period to use, if times are periodic. If supplied, input times
+        must be sorted according to t % period!
+    subtract_mid : boolean
+        If true, then subtract the middle value from each sum
+
+    Returns
+    -------
+    arrays : tuple of ndarrays
+        arrays containing the windowed sum of each input array
+    """
+    span = np.asarray(span, dtype=int)
+    if not np.all(span > 0):
+        raise ValueError("span values must be positive")
+    
+    arrays = tuple(map(np.asarray, arrays))
+    N = arrays[0].size
+    if not all(a.shape == (N,) for a in arrays):
+        raise ValueError("sizes of provided arrays must match")
+    
+    t_input = t
+    if t is not None:
+        t = np.asarray(t)
+        if t.shape != (N,):
+            raise ValueError("shape of t must match shape of arrays "
+                             "t -> {0} arr -> {1}".format(t.shape,
+                                                          arrays[0].shape))
+    else:
+        # XXX: special-case no t?
+        t = np.ones(N)
+    
+    tpowers = tpowers + np.zeros(len(arrays))
+    if len(tpowers) != len(arrays):
+        raise ValueError("tpowers must be broadcastable with number of arrays")
+
+    if period:
+        if t_input is None:
+            raise ValueError("periodic requires t to be provided")
+        t = t % period
+    
     if indices is not None:
         span, indices = np.broadcast_arrays(span, indices)
 
-    if slow:
-        # Slow fixed/variable span
-        results = []
-        for array in map(np.asarray, arrays):
-            assert array.ndim == 1
-            if indices is None:
-                span, array = np.broadcast_arrays(span, array)
-                ind_spans = enumerate(span)
-                count = len(span)
+    # For the periodic case, re-call the function with padded arrays
+    if period:
+        t, arrays, sl = _pad_arrays(t, arrays, indices, span, period)
+        if len(t) == N:
+            # No padding needed! We'll carry-on with the algorithm here.
+            pass
+        else:
+            # arrays are padded. Recursively call windowed_sum() and return.
+            if span.ndim == 0 and indices is None:
+                # fixed-span/no index case is done faster this way
+                arrs = windowed_sum(arrays, span, t=t, indices=indices,
+                                    tpowers=tpowers, period=None,
+                                    subtract_mid=subtract_mid)
+                return tuple([a[sl] for a in arrs])
             else:
-                ind_spans = np.broadcast(indices, span)
-                count = len(indices)
-            result = (array[max(0, i - s // 2): i - s // 2 + s].sum()
-                      for i, s in ind_spans)
-            results.append(np.fromiter(result, dtype=array.dtype, count=count))
+                # this works for variable span and general indices
+                if indices is None:
+                    indices = np.arange(N)
+                indices = indices + sl.start
+                return windowed_sum(arrays, span, t=t, indices=indices,
+                                    tpowers=tpowers, period=None,
+                                    subtract_mid=subtract_mid)
 
-    elif span.ndim == 0:
-        # Fast fixed-span
-        window = np.ones(span)        
-        def convolve_same(a, window):
-            res = np.convolve(a, window, mode='same')
-            if len(res) > len(a):
-                res = res[len(res) // 2 - len(a) // 2:][:len(a)]
-            return res
-        results = [convolve_same(a, window) for a in arrays]
+    # Now we should be safely in the non-periodic case.
+    if span.ndim == 0:
+        # fixed-span case
+        window = np.ones(span)
+        if period:
+            # This should have been taken care of above
+            raise ValueError("Something went terribly wrong...")
+        else:
+            def convolve_same(a, window):
+                if len(window) <= len(a):
+                    res = np.convolve(a, window, mode='same')
+                else:
+                    res = np.convolve(a, window, mode='full')
+                    start = (len(window) - 1) // 2
+                    res = res[start:start + len(a)]
+                return res
+            results = [convolve_same(a * t ** tp, window)
+                       for a, tp in zip(arrays, tpowers)]
+            indices = slice(None) # for below
 
     else:
-        # Fast variable-span
+        # variable-span case
         if indices is None:
             indices = np.arange(len(span))
-        mins = np.asarray(indices) - span // 2
+        if period:
+            # This should have been taken care of above
+            raise ValueError("Something went terribly wrong...")
+        else:
+            mins = np.asarray(indices) - span // 2
+            results = []
+            for a, tp in zip(arrays, tpowers):
+                ranges = np.vstack([np.maximum(0, mins),
+                                    np.minimum(len(a), mins+span)]).ravel('F')
+                results.append(np.add.reduceat(np.append(a * t ** tp, 0),
+                                               ranges)[::2])
 
-        results = []
-        for a in map(np.asarray, arrays):
-            ranges = np.vstack([np.maximum(0, mins),
-                                np.minimum(len(a), mins + span)]).ravel('F')
-            results.append(np.add.reduceat(np.append(a, 0), ranges)[::2])
-
+    # Subtract the midpoint if required: this is used in cross-validation
     if subtract_mid:
-        if indices is None:
-            indices = slice(None)
-        results = (r - a[indices] for r, a in zip(results, arrays))
+        results = [r - a[indices] * t[indices] ** tp
+                   for r, a, tp in zip(results, arrays, tpowers)]
 
     return tuple(results)
 
 
+def _pad_arrays(t, arrays, indices, span, period):
+    """Internal routine to pad arrays for periodic models."""
+    N = len(t)
+
+    if indices is None:
+        indices = np.arange(N)
+    pad_left = max(0, 0 - np.min(indices - span // 2))
+    pad_right = max(0, np.max(indices + span - span // 2) - (N - 1))
+
+    if pad_left + pad_right > 0:
+        Nright, pad_right = divmod(pad_right, N)
+        Nleft, pad_left = divmod(pad_left, N)
+        t = np.concatenate([t[N - pad_left:] - (Nleft + 1) * period]
+                           + [t + i * period
+                              for i in range(-Nleft, Nright + 1)]
+                           + [t[:pad_right] + (Nright + 1) * period])
+        arrays = [np.concatenate([a[N - pad_left:]]
+                                 + (Nleft + Nright + 1) * [a]
+                                 + [a[:pad_right]])
+                  for a in arrays]
+        pad_left = pad_left % N
+        Nright = pad_right / N
+        pad_right = pad_right % N
+
+        return (t, arrays, slice(pad_left + Nleft * N,
+                                 pad_left + (Nleft + 1) * N))
+    else:
+        return (t, arrays, slice(None))
+
+
+def _prep_smooth(t, y, dy, span, t_out, span_out, period):
+    """Private function to prepare & check variables for smooth utilities"""
+
+    # If period is provided, sort by phases. Otherwise sort by t
+    if period:
+        t = t % period
+        if t_out is not None:
+            t_out = t_out % period
+
+    t, y, dy = validate_inputs(t, y, dy, sort_by=t)
+
+    if span_out is not None:
+        if t_out is None:
+            raise ValueError("Must specify t_out when span_out is given")
+        if span is not None:
+            raise ValueError("Must specify only one of span, span_out")
+        span, t_out = np.broadcast_arrays(span_out, t_out)
+        indices = np.searchsorted(t, t_out)
+    elif span is None:
+        raise ValueError("Must specify either span_out or span")
+    else:
+        indices = None
+
+    return t, y, dy, span, t_out, span_out, indices
+
+
 def moving_average_smooth(t, y, dy, span=None, cv=True,
-                          t_out=None, span_out=None):
+                          t_out=None, span_out=None, period=None):
     """Perform a moving-average smooth of the data
 
     Parameters
@@ -176,28 +360,19 @@ def moving_average_smooth(t, y, dy, span=None, cv=True,
         the output times for the moving averages
     span_out : array_like (optional)
         the spans associated with the output times t_out
+    period : float
+        if provided, then consider the inputs periodic with the given period
 
     Returns
     -------
     y_smooth : array_like
         smoothed y values at each time t (or t_out)
     """
-    t, y, dy = validate_inputs(t, y, dy, sort_by=t)
-
-    if span_out is not None:
-        if t_out is None:
-            raise ValueError("Must specify t_out when span_out is given")
-        if span is not None:
-            raise ValueError("Must specify only one of span, span_out")
-        span, t_out = np.broadcast_arrays(span_out, t_out)
-        indices = np.searchsorted(t, t_out)
-    elif span is None:
-        raise ValueError("Must specify either span_out or span")
-    else:
-        indices = None
+    prep = _prep_smooth(t, y, dy, span, t_out, span_out, period)
+    t, y, dy, span, t_out, span_out, indices = prep
 
     w = dy ** -2
-    w, yw = windowed_sum(w, y * w, span=span, subtract_mid=cv,
+    w, yw = windowed_sum([w, y * w], span=span, subtract_mid=cv,
                          indices=indices)
 
     if t_out is None or span_out is not None:
@@ -208,7 +383,7 @@ def moving_average_smooth(t, y, dy, span=None, cv=True,
 
 
 def linear_smooth(t, y, dy, span=None, cv=True,
-                  t_out=None, span_out=None):
+                  t_out=None, span_out=None, period=None):
     """Perform a linear smooth of the data
 
     Parameters
@@ -224,6 +399,8 @@ def linear_smooth(t, y, dy, span=None, cv=True,
         the output times for the moving averages
     span_out : array_like (optional)
         the spans associated with the output times t_out
+    period : float
+        if provided, then consider the inputs periodic with the given period
 
     Returns
     -------
@@ -231,22 +408,12 @@ def linear_smooth(t, y, dy, span=None, cv=True,
         smoothed y values at each time t or t_out
     """
     t_input = t
-    t, y, dy = validate_inputs(t, y, dy, sort_by=t)
-
-    if span_out is not None:
-        if t_out is None:
-            raise ValueError("Must specify t_out when span_out is given")
-        if span is not None:
-            raise ValueError("Must specify only one of span, span_out")
-        span, t_out = np.broadcast_arrays(span_out, t_out)
-        indices = np.searchsorted(t, t_out)
-    elif span is None:
-        raise ValueError("Must specify either span_out or span")
-    else:
-        indices = None
+    prep = _prep_smooth(t, y, dy, span, t_out, span_out, period)
+    t, y, dy, span, t_out, span_out, indices = prep
 
     w = dy ** -2
-    w, tw, yw, ttw, tyw = windowed_sum(w, t * w, y * w, t * t * w, t * y * w,
+    w, yw, tw, tyw, ttw = windowed_sum([w, y * w, w, y * w, w], t=t,
+                                       tpowers=[0, 0, 1, 1, 2],
                                        span=span, indices=indices,
                                        subtract_mid=cv)
 
@@ -304,3 +471,5 @@ def multinterp(x, y, xquery, slow=False):
         j = np.arange(len(xquery))
         return y[i, j] + ((xquery - x[i]) *
                           (y[i + 1, j] - y[i, j]) / (x[i + 1] - x[i]))
+
+
